grl: fix compilation errors and issues
Warning, mostly regarding possible type mismatch, were fixed. Static
casts were added, where needed, to satyisfy the compiler.

Versioning of the lib was changed, and now it will be generation v0.2,
as currently 3rd release before the v1.0 release is being developed.

There were 2 errors in the debug release, which were fixed:
  * Output of the project was changed, so it would generate static lib
  * Pointer could not be initialized in one assignment (in descriptor)

+++ OpenGRL/include/grl/gesture/FloodFillClipped.h
@@ -104,13 +104,13 @@ VoxelArray2D::getData()
 class FloodFillClipped
 {
 public:
-	void setup(double tolerance, const cv::Mat &source);
+	void setup(int tolerance, const cv::Mat &source);
 
 	bool extractObject(Vec2i startingPoint, Object &object, Plane plane);
 	bool isAnalyzed(Vec2i point);
 
 private:
-	double _tolerance;
+	int _tolerance;
 	VoxelArray2D _voxelImage;
 
 };
@@ -122,7 +122,7 @@ FloodFillClipped::isAnalyzed(Vec2i point)
 }
 
 inline void
-FloodFillClipped::setup(double tolerance, const cv::Mat &source)
+FloodFillClipped::setup(int tolerance, const cv::Mat &source)
 {
 	_tolerance = tolerance;
 	_voxelImage.fromImage(source, tolerance);
